+++ lib/ratatui_ruby/schema/paragraph.rb
@@ -27,7 +27,7 @@ module RatatuiRuby
     #
     #   # Scrolling mechanism
     #   Paragraph.new(text: large_text, scroll: [scroll_y, 0])
-    class Paragraph < Data.define(:text, :style, :block, :wrap, :align, :scroll)
+    class Paragraph < Data.define(:text, :style, :block, :wrap, :alignment, :scroll)
       ##
       # :attr_reader: text
       # The content to display.
@@ -45,7 +45,7 @@ module RatatuiRuby
       # Whether to wrap text at the edge of the container (Boolean).
 
       ##
-      # :attr_reader: align
+      # :attr_reader: alignment
       # Text alignment.
       #
       # <tt>:left</tt>, <tt>:center</tt>, or <tt>:right</tt>.
@@ -60,24 +60,24 @@ module RatatuiRuby
       # [style] Style object.
       # [block] Block object.
       # [wrap] Boolean (default: false).
-      # [align] Symbol (default: <tt>:left</tt>).
+      # [alignment] Symbol (default: <tt>:left</tt>).
       # [scroll] Array of [y, x] integers (duck-typed via +to_int+).
-      def initialize(text:, style: Style.default, block: nil, wrap: false, align: :left, scroll: [0, 0])
+      def initialize(text:, style: Style.default, block: nil, wrap: false, alignment: :left, scroll: [0, 0])
         super(
           text: text,
           style: style,
           block: block,
           wrap: wrap,
-          align: align,
+          alignment: alignment,
           scroll: [Integer(scroll[0]), Integer(scroll[1])]
         )
       end
 
       # Legacy constructor support.
-      def self.new(text:, style: nil, fg: nil, bg: nil, block: nil, wrap: false, align: :left, scroll: [0, 0])
+      def self.new(text:, style: nil, fg: nil, bg: nil, block: nil, wrap: false, alignment: :left, scroll: [0, 0])
         style ||= Style.new(fg:, bg:)
         coerced_scroll = [Integer(scroll[0]), Integer(scroll[1])]
-        super(text:, style:, block:, wrap:, align:, scroll: coerced_scroll)
+        super(text:, style:, block:, wrap:, alignment:, scroll: coerced_scroll)
       end
 
       # Returns the number of lines the paragraph would take up if rendered with the given width.